+++ src/main/types.ts
@@ -324,6 +324,7 @@ enum Pages {
   'None',
   'SceneEditor',
   'Settings',
+  'InstantReplay',
 }
 
 /**