+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 ### Added
+- Added the instant replay feature. You can now view a recording, as it is recording, without interrupting it.
 ### Fixed
 
 ## [7.9.0] - 2026-06-26